requirements.txt
#ipykernel
scikit-build
pyvista
pyikarus
pandas
tabulate
